pyproject.toml
[tool.poetry]
name = "turpoints"
version = "0.1.0"
description = ""
authors = ["61246880 <jorge.plautz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
djangorestframework = "^3.12.2"
gunicorn = "^20.0.4"
python-decouple = "^3.4"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"
django-s3-folder-storage = "^0.5"
Collectfast = "^2.2.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest-django = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
